requirements.txt
laplace_torch==0.1a2
matplotlib==3.7.1
numpy==1.24.3
scipy==1.11.3
seaborn==0.13.0
streamlit==1.25.0
tensorflow==2.13.0
tensorflow_macos==2.13.0
torch==2.0.0
pandas==2.0.1
scikit_learn==1.2.2